refactor: change status_code data type

+++ database.sql
@@ -7,7 +7,7 @@ CREATE TABLE IF NOT EXISTS urls (
 CREATE TABLE IF NOT EXISTS url_checks (
     id bigint PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     url_id bigint REFERENCES urls (id),
-    status_code numeric,
+    status_code integer,
     h1 varchar(255),
     title varchar(255),
     description text,
